Split the command parser
